[Refactor] PlayerType::town_num をAngbandWorld に移した

+++ src/system/floor/town-list.h
@@ -4,7 +4,7 @@
 #include "util/abstract-vector-wrapper.h"
 #include <vector>
 
-constexpr short VALID_TOWNS = 6; // @details 旧海底都市クエストのマップを除外する. 有効な町に差し替え完了したら不要になるので注意.
+constexpr size_t VALID_TOWNS = 6; // @details 旧海底都市クエストのマップを除外する. 有効な町に差し替え完了したら不要になるので注意.
 
 class TownList : public util::AbstractVectorWrapper<TownInfo> {
 public:
